Seperate lib and main
